+++ src/main/java/br/com/marmitaria/dto/carrinho/RespostaCarrinhoDTO.java
@@ -1,12 +1,14 @@
 package br.com.marmitaria.dto.carrinho;
 
+import br.com.marmitaria.dto.item.RespostaItemDTO;
+
 import java.math.BigDecimal;
 import java.util.List;
 
 public record RespostaCarrinhoDTO(
         Long id,
         Long usuarioId,
-        List<RespostaCarrinhoItemDTO> itens,
+        List<RespostaItemDTO> itens,
         int totalProdutos,
         BigDecimal valorTotal
 ) {}
